Models are not a service, move it into shared folder as official style guide says
