+++ client/app/account/account.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ToastComponent } from '../shared/toast/toast.component';
 import { AuthService } from '../services/auth.service';
 import { UserService } from '../services/user.service';
-import { User } from '../services/models/user';
+import { User } from '../shared/models/user.model';
 
 @Component({
   selector: 'app-account',
