+++ client/app/services/cat.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
-import { Cat } from './models/cat';
+
+import { Cat } from '../shared/models/cat.model';
 
 @Injectable()
 export class CatService {
