+++ client/app/shared/models/cat.model.ts
